review

+++ src/com/zimcarry/review/ReviewDTO.java
@@ -4,11 +4,11 @@ import java.util.Date;
 
 public class ReviewDTO {
 	private long reIdx;
-	private int reScore;
+	private double reScore;
 	private String reTitle;
 	private String reContent;
 	private String reWriter;
 	private Date reWritedate;
 	private long reBookIdx;
-	
+
 }
